refactor checkout modules into features and adapters
