+++ storefronts/adapters/core/checkout-adapter.js
@@ -1,0 +1,3 @@
+export * from '../../features/checkout/checkout-core.js';
+import * as checkout from '../../features/checkout/checkout-core.js';
+export default checkout;
