+++ storefronts/features/checkout/utils/authorizeNetIframeStyles.js
