+++ storefronts/features/checkout/utils/cartHash.js
