+++ storefronts/features/checkout/utils/resolveGateway.js
@@ -1,4 +1,4 @@
-import resolveGateway from '../../../core/utils/resolveGateway.js';
+import resolveGateway from '../../../../core/utils/resolveGateway.js';
 
 export default function getActivePaymentGateway(log = () => {}, warn = () => {}) {
   const cfg = window.SMOOTHR_CONFIG || {};
