--- storefronts/platforms/core/checkout.js
@@ -1,3 +1,0 @@
-export * from '../../core/checkout.js';
-import * as checkout from '../../core/checkout.js';
-export default checkout;
