+++ storefronts/tests/platforms/webflow/checkout.dom.test.ts
@@ -29,12 +29,12 @@ afterEach(() => {
 describe('webflow checkout adapter dom', () => {
 
   it('loads without modifying global flag', async () => {
-    await import('../../../platforms/webflow/checkout.js');
+    await import('../../../adapters/webflow/initCheckoutWebflow.js');
     expect((window as any).__SMOOTHR_CHECKOUT_INITIALIZED__).toBeUndefined();
 
     vi.resetModules();
     (window as any).__SMOOTHR_CHECKOUT_INITIALIZED__ = true;
-    await import('../../../platforms/webflow/checkout.js');
+    await import('../../../adapters/webflow/initCheckoutWebflow.js');
     expect((window as any).__SMOOTHR_CHECKOUT_INITIALIZED__).toBe(true);
   });
